feat: http caching (#47)
This implements an HTTP client that will (mostly) respect
`Cache-Control` headers.

Architecture of the cache:

 * Cache is persisted to discuss in a JSON file
* Deserialization does not seem to cause noticeable delay on startup
(tested with ~6MB cache)
* Serialization blocks esbuild (cache saving is done at the end of the
esbuild pipeline), but the overall perf was unaffected even when this
was done concurrently/skipped
* Cache contains byte size, and a hashmap of URLs -> entries. Entries
store a deadline (or nil for immutable), contents as string (esbuild
takes a string pointer rather than a byte slice, so it was faster to
just store as a string), and an int64ptr that holds the time of last
use.
* Retrieving a value from the cache causes the time of last use to be
atomically updated (not fully atomic, since we generate and then write,
but worst case, this causes a few nanosecs diff)
 * On retrieve, we delete the entry if it has passed its TTL
* On save/flush, the cache performs a 'compaction' cycle, where it
evicts all expired data and then finally deletes from the least recently
used items until the cache reaches the desired size. It then flushes the
cache to disk.

---

It significantly helps with #46, though does not fix it completely.

 * On main, cold loading `@robinplatform/app-example` takes around 22s.
   * 16.5s to render `base.html`
* On httpcache, cold loading `@robinplatform/app-example` takes around
0.3s.
   * 150ms to render `base.html`
* On either, cold loading `@robinplatform/app-example` from disk takes
around 0.14s. (how on earth is this a different order of
magnitude.........)
   * 70ms to render `base.html`

---------

Co-authored-by: Albert Liu <albertymliu@gmail.com>