feat: apps directory (#62)
This PR adds an `apps` directory to the project directory (i.e.
`~/.robin/projects/[project]/apps/[app]`), where the built apps live and
run from. Since this is now the working directory for the app, the
project directory is passed as an environment variable
(`ROBIN_PROJECT_PATH`) to the app.

We can use this as the bundle cache for #54, but it also makes more
sense for the app to work out of its own directory (to store temporary
files, assets, etc), especially for #51.