Allow for synchronization between pubsub and something else (#117)
Allow for pubsub messages to be synchronized to a point in time via a
counter on the topic. Since topics are synchronized with a mutex, you
can be sure that if you call `topic.GetInfo()` and get back
`info.Counter = 19`, that all messages with a `MessageId` less than 20
have already sent, and that 20 will be the message ID of the next
message that is sent.

This property can be used to synchronize state with messages which act
as modifications on the state.

1. Subscribe to the state value, so that you don't miss any messages
2. Fetch the state value, with a counter
3. Throw away any messages received with a lower counter value than the
state
4. Apply all remaining messages to the state's local copy

This PR also does the following:
- implements pubsub for apps, so an app daemon can publish to a topic
and receive information from a topic.
- Removes the project name from the topic ID, and process ID, because
its unnecessary. I've also removed it from process ID, although that
technically breaks something if there's multiple projects. However, for
now we can treat that as not happening meaningfully; additionally, I've
added a TODO to fix it.
- Adds a usage of the app daemon pubsub to the POGO app; its very
simple, and also very finnicky, but it does at least mostly do its job.