Allow `CreateTopic` RPC call to succeed when topic already exists (#120)
Right now, the topics associated with an app don't die when the app
dies. To prevent the crash that this problem causes, this PR implements
the band-aid of just making `CreateTopic` succeed when the topic already
exists. However, we still ultimately gotta fix these band-aids
eventually, because this band-aid exists because the app lifetime is
difficult to query properly, because of ANOTHER bandaid.