Delete POGO App (#132)
What I've found during POGO App dev:
- Restarting servers, either Robin or app servers, needs to be done
manually
- Build/refresh times were kinda slow, ~600ms
- Memory usage (both browser and robin) sometimes seemed to cause my
computer to have trouble
- Sometimes the page gets hung and I can't even refresh the page. I have
no idea what this is or why it happens. My hunch is it has to do with
requests getting hung when re-opening the page.
- Some data in JS isn't serializable, which makes using the RPC handlers
a little funky. Most notably, `Date` objects get serialized to strings,
and `undefined` can't be serialized, so RPC mutations need to always
return an empty object in place of `void`